Fix PublishItemsOutputGroup computation
This change fixes a failure in merging release/5.0.1xx-preview1 to master

Given 1f2a4a1, we should undo most of the change in 43eb71d, because `ResolvedFilesToPublish` is actually consistent now in both single-file and non-single-file scenarios.
In `PublishSingleFile` case, we only need to add the single-file bundle itself.
No special handling is necessary for the deps.json file, since it is now included within `ResolvedFilesToPublish` list.